backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
email-validator